requirements.txt
# tensorflow==2.20.0
tensorflow==2.15.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy
fastapi[standard]
requests
tqdm